requirements.txt
boto3
pandas
sqlalchemy
python-dotenv
pymysql